Apply CR comments

+++ onnx/backend/test/case/node/bernoulli.py
@@ -26,7 +26,7 @@ class Bernoulli(Base):
         )
 
         x = np.random.uniform(0.0, 1.0, 10).astype(np.float32)
-        y = bernoulli_reference_implementation(x, np.float32)
+        y = bernoulli_reference_implementation(x, float)
         expect(node, inputs=[x], outputs=[y], name="test_bernoulli")
 
     @staticmethod
@@ -39,7 +39,7 @@ class Bernoulli(Base):
         )
 
         x = np.random.uniform(0.0, 1.0, 10).astype(np.float32)
-        y = bernoulli_reference_implementation(x, np.float64)
+        y = bernoulli_reference_implementation(x, float)
         expect(node, inputs=[x], outputs=[y], name="test_bernoulli_double")
 
     @staticmethod
@@ -53,5 +53,5 @@ class Bernoulli(Base):
         )
 
         x = np.random.uniform(0.0, 1.0, 10).astype(np.float32)
-        y = bernoulli_reference_implementation(x, np.float32)
+        y = bernoulli_reference_implementation(x, float)
         expect(node, inputs=[x], outputs=[y], name="test_bernoulli_seed")
